Refactor moving the log_performance context to a different submodule (#604)
* Refactor moving the log_performance context to a different submodule

* Removed unused imports
